Refactor leading semicolon print
